# docs/swagger.yaml
basePath: /
definitions:
...
  domain.CategoryStat:
    properties:
      name:
        type: string
      share:
        description: Percentage of total (optional, can be calculated on frontend)
        type: number
      value:
        type: integer
    type: object
  domain.GeoStat:
    properties:
      country:
        type: string
      value:
        type: integer
    type: object
  domain.HeatmapPoint:
    properties:
      count:
        description: Color intensity
        type: integer
      day:
        description: 0=Sun, 1=Mon, ..., 6=Sat
        type: integer
      hour:
        description: 0-23
        type: integer
    type: object
...
  domain.SankeyData:
    properties:
      links:
        items:
          $ref: '#/definitions/domain.SankeyLink'
        type: array
      nodes:
        items:
          $ref: '#/definitions/domain.SankeyNode'
        type: array
    type: object
  domain.SankeyLink:
    properties:
      source:
        type: string
      target:
        type: string
      value:
        type: integer
    type: object
  domain.SankeyNode:
    properties:
      id:
        description: Node name (e.g., "USA" or "Mobile")
        type: string
      layer:
        description: Column (0 - Referer, 1 - Device...)
        type: integer
    type: object
  domain.Summary:
    properties:
      top_browsers:
        items:
          $ref: '#/definitions/domain.CategoryStat'
        type: array
      top_os:
        items:
          $ref: '#/definitions/domain.CategoryStat'
        type: array
      total_clicks:
        type: integer
    type: object
  domain.TrafficQuality:
    properties:
      bot_score:
        type: integer
      geo_diversity_score:
        type: integer
      human_score:
        type: integer
      is_suspicious:
        type: boolean
      mobile_friendly_score:
        type: integer
    type: object
  domain.User:
    properties:
      clicks_current_month:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      plan_expires:
        description: nil = forever (for free)
        type: string
      plan_id:
        type: string
      role:
        $ref: '#/definitions/domain.UserRole'
    type: object
  domain.UserRole:
    enum:
    - admin
    - client
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleClient
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  v1.AnalyticsSummaryResponse:
    properties:
      data:
        $ref: '#/definitions/domain.Summary'
    type: object
...
  v1.GeoResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.GeoStat'
        type: array
    type: object
  v1.HeatmapResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.HeatmapPoint'
        type: array
    type: object
...
  v1.LoginData:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  v1.LoginReq:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret123
        type: string
    required:
    - email
    - password
    type: object
  v1.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/v1.LoginData'
    type: object
...
  v1.QualityResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.TrafficQuality'
        type: array
    type: object
  v1.RegisterReq:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret123
        type: string
    required:
    - email
    - password
    type: object
  v1.RegisterResponse:
    properties:
      data:
        $ref: '#/definitions/domain.User'
    type: object
  v1.SankeyResponse:
    properties:
      data:
        $ref: '#/definitions/domain.SankeyData'
    type: object
  v1.StatsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.CategoryStat'
        type: array
    type: object
...
  v1.StreamGraphResponse:
    properties:
      data:
        description: Предполагаем тип возвращаемого значения сервиса
        items:
          additionalProperties: {}
          type: object
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Traffic analytics service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Traffic Analytics
  version: "1.0"
paths:
...
  /api/v1/analytics/flow:
    get:
      description: Get flow data (Referer -> Device -> Country)
      parameters:
      - description: Filter by Campaign
        in: query
        name: campaign_id
        type: string
      - description: Filter by Link
        in: query
        name: link_id
        type: string
      - description: Date From (RFC3339)
        in: query
        name: from
        type: string
      - description: Date To (RFC3339)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SankeyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Sankey flow
      tags:
      - Analytics
  /api/v1/analytics/geo:
    get:
      description: Get clicks count by country
      parameters:
      - description: Filter by Campaign
        in: query
        name: campaign_id
        type: string
      - description: Filter by Link
        in: query
        name: link_id
        type: string
      - description: Date From (RFC3339)
        in: query
        name: from
        type: string
      - description: Date To (RFC3339)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GeoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Geo distribution
      tags:
      - Analytics
  /api/v1/analytics/heatmap:
    get:
      description: Get click intensity by Day of Week and Hour
      parameters:
      - description: Filter
        in: query
        name: campaign_id
        type: string
      - description: Filter
        in: query
        name: link_id
        type: string
      - description: Date From (RFC3339)
        in: query
        name: from
        type: string
      - description: Date To (RFC3339)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.HeatmapResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Heatmap data
      tags:
      - Analytics
  /api/v1/analytics/quality:
    get:
      description: Get quality metrics for Radar Chart
      parameters:
      - description: Filter by Campaign
        in: query
        name: campaign_id
        type: string
      - description: Filter by Link
        in: query
        name: link_id
        type: string
      - description: Date From (RFC3339)
        in: query
        name: from
        type: string
      - description: Date To (RFC3339)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.QualityResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Traffic quality
      tags:
      - Analytics
  /api/v1/analytics/stats:
    get:
      description: Get top metrics for a dimension (browser, os, device)
      parameters:
      - default: browser
        description: browser, os, device
        in: query
        name: dimension
        type: string
      - description: Filter by Campaign
        in: query
        name: campaign_id
        type: string
      - description: Filter by Link
        in: query
        name: link_id
        type: string
      - description: Date From (RFC3339)
        in: query
        name: from
        type: string
      - description: Date To (RFC3339)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.StatsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Category Stats (Pie/Bar)
      tags:
      - Analytics
  /api/v1/analytics/stream:
    get:
      description: Get time-series data for Streamgraph/Stacked Area charts
      parameters:
      - default: os
        description: os, browser, country
        in: query
        name: group_by
        type: string
      - description: Filter by Campaign
        in: query
        name: campaign_id
        type: string
      - description: Filter by Link
        in: query
        name: link_id
        type: string
      - description: Date From (RFC3339)
        in: query
        name: from
        type: string
      - description: Date To (RFC3339)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.StreamGraphResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Stream graph data
      tags:
      - Analytics
  /api/v1/analytics/summary:
    get:
      description: Get aggregated stats (clicks, uniques, etc.)
      parameters:
      - description: Filter by Campaign
        in: query
        name: campaign_id
        type: string
      - description: Filter by Link
        in: query
        name: link_id
        type: string
      - description: Date From (RFC3339)
        in: query
        name: from
        type: string
      - description: Date To (RFC3339)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.AnalyticsSummaryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Analytics summary
      tags:
      - Analytics
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate and get JWT token
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Token and User info
          schema:
            $ref: '#/definitions/v1.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created user data
          schema:
            $ref: '#/definitions/v1.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Missing authorization header
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Email already taken
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register new user
      tags:
      - Auth
...
securityDefinitions:
  APIKeyAuth:
    description: Static API Key for admin/service access
    in: header
    name: X-API-Key
    type: apiKey
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"